requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0
factory-boy==2.12.0
Faker==3.0.0
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
responses==0.10.9
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
wcwidth==0.1.8
zipp==0.6.0
